Refactor package structure
And unify copyright headers on java files.
