+++ vaadin-cdi/src/main/java/com/vaadin/cdi/VaadinExtension.java
@@ -1,12 +1,12 @@
 /*
- * Copyright 2000-2013 Vaadin Ltd.
- * 
+ * Copyright 2000-2018 Vaadin Ltd.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,10 +14,15 @@
  * the License.
  */
 
-package com.vaadin.cdi.internal;
+package com.vaadin.cdi;
 
-import com.vaadin.cdi.*;
-import com.vaadin.cdi.internal.InconsistentDeploymentException.ID;
+import com.vaadin.cdi.InconsistentDeploymentException.ID;
+import com.vaadin.cdi.annotation.NormalUIScoped;
+import com.vaadin.cdi.annotation.UIScoped;
+import com.vaadin.cdi.annotation.VaadinSessionScoped;
+import com.vaadin.cdi.context.ContextWrapper;
+import com.vaadin.cdi.context.UIScopedContext;
+import com.vaadin.cdi.context.VaadinSessionScopedContext;
 import com.vaadin.flow.component.Component;
 
 import javax.enterprise.event.Observes;
