+++ vaadin-cdi/src/main/java/com/vaadin/cdi/context/VaadinSessionScopedContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2013 Vaadin Ltd.
+ * Copyright 2000-2018 Vaadin Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -12,12 +12,11 @@
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations under
  * the License.
- *
  */
 
-package com.vaadin.cdi.internal;
+package com.vaadin.cdi.context;
 
-import com.vaadin.cdi.VaadinSessionScoped;
+import com.vaadin.cdi.annotation.VaadinSessionScoped;
 import com.vaadin.flow.server.VaadinSession;
 import org.apache.deltaspike.core.util.ContextUtils;
 import org.apache.deltaspike.core.util.context.AbstractContext;
